% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNA-checks.R
\docType{methods}
\name{checktRNAGRanges}
\alias{checktRNAGRanges}
\alias{checktRNAGRanges}
\alias{checktRNAGRanges,GRanges-method}
\title{tRNA compatibility check}
\usage{
checktRNAGRanges(gr)

\S4method{checktRNAGRanges}{GRanges}(gr)
}
\arguments{
\item{gr}{the \code{GRanges} object to test}
}
\value{
a logical value
}
\description{
\code{checktRNAGRanges} checks whether a GRanges object contains the
information expected for a tRNA result.
}
\examples{
data("gr", package = "tRNA", envir = environment())
checktRNAGRanges(gr)
}
