% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNA-dotbracket.R
\docType{methods}
\name{gettRNABasePairing}
\alias{gettRNABasePairing}
\alias{gettRNALoopIDs}
\alias{getBasePairing}
\alias{gettRNABasePairing}
\alias{gettRNALoopIDs}
\alias{getLoopIDs}
\alias{gettRNABasePairing,GRanges-method}
\alias{getBasePairing}
\alias{gettRNALoopIDs,GRanges-method}
\alias{getLoopIDs}
\title{Accessing Dot Bracket annotation}
\usage{
gettRNABasePairing(gr)

gettRNALoopIDs(gr)

\S4method{gettRNABasePairing}{GRanges}(gr)

getBasePairing(dotBracket, sequence)

\S4method{gettRNALoopIDs}{GRanges}(gr)

getLoopIDs(dotBracket)
}
\arguments{
\item{gr}{a GRanges object created by \code{import.tRNAscanAsGRanges} or
GRanges with equivalent information. The \code{tRNA_str} column will be used
for input into \code{getBasePairing}.}

\item{dotBracket}{character vectors describing a nucleotide sequence
structure in the dot bracket annotations. Valid characters are:
\code{.(\\{[><]\\})}}

\item{sequence}{optional: character vectors describing a nucleotide sequence.
The same number of sequences with the same length as the dot bracket string 
have to be used. Each nucleotide sequence has to be a character vector.
The identity of the nucleotides are not control, so in theory all letters 
can be used.}
}
\value{
\code{getBasePairing}: 
The result is a data.frame with three columns: pos, forward, reverse. If a
position is unpaired, forward and reverse will be \code{0}, otherwise it
will match the base paired positions.
\code{getHairpinLoops}, \code{getStems}: 
return a list of list of \code{IRanges} describing all coordinates found for
the given structures.
}
\description{
\code{getBasePairing} converts a dot bracket annotation into a
\code{data.frame}. Base pairing is indicated by corresponding numbers
in the forward and reverse columns.

\code{getHairpinLoops} and \code{getStems} are helper functions for getting
coordinates for stems and hairpin loops as \code{IRanges} objects.
}
