% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNA-features.R
\docType{methods}
\name{gettRNASummary}
\alias{gettRNASummary}
\alias{gettRNASummary}
\alias{gettRNASummary,GRanges-method}
\title{Summary of tRNA features}
\usage{
gettRNASummary(gr)

\S4method{gettRNASummary}{GRanges}(gr)
}
\arguments{
\item{gr}{a GRanges object, which passes the \code{istRNAGRanges} test.}

\item{grl}{a GRangesList object with GRanges elements , which passes the 
\code{istRNAGRanges} test.}
}
\value{
a DataFrame object
}
\description{
\code{gettRNASummary} prepares a DataFrame with the aggregated features
of tRNAs from a GRanges object. Logical values are converted to numeric
values.
}
\examples{
data("gr", package = "tRNA", envir = environment())
gettRNASummary(gr)
}
