% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNA-checks.R
\docType{methods}
\name{istRNAGRanges}
\alias{istRNAGRanges}
\alias{istRNAGRanges}
\alias{istRNAGRanges,GRanges-method}
\title{tRNA compatibility check}
\usage{
istRNAGRanges(x)

\S4method{istRNAGRanges}{GRanges}(x)
}
\arguments{
\item{x}{the \code{GRanges} object to test}
}
\value{
a logical value
}
\description{
\code{istRNAGRanges} checks whether a GRanges object contains the
information expected for a tRNA result.
}
\examples{
data("gr", package = "tRNA", envir = environment())
istRNAGRanges(gr)
}
