% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNA-plot.R
\docType{data}
\name{plottRNAFeatures}
\alias{plottRNAFeatures}
\alias{gettRNAFeaturePlots}
\alias{plottRNAFeatures}
\alias{TRNA_PLOT_LABELS}
\alias{gettRNAFeaturePlots,GRangesList-method}
\title{Graphical summary of tRNA features}
\format{An object of class \code{list} of length 40.}
\usage{
plottRNAFeatures(grl)

gettRNAFeaturePlots(grl)

TRNA_PLOT_LABELS

\S4method{gettRNAFeaturePlots}{GRangesList}(grl)
}
\value{
a ggplot2
}
\description{

}
\examples{
\donttest{
}
}
\keyword{datasets}
