% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNA.R
\docType{data}
\name{tRNA-data}
\alias{tRNA-data}
\alias{gr}
\title{tRNA example data}
\format{object of class \code{GRanges}}
\usage{
tRNA
}
\description{
Example data for using the tRNA package
}
\keyword{datasets}
\keyword{tRNA}
