% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNA-features.R
\name{gettRNASummary}
\alias{gettRNASummary}
\alias{gettRNASummary,GRangesList-method}
\alias{gettRNASummary,GRanges-method}
\title{Summary of tRNA features}
\usage{
gettRNASummary(x)

\S4method{gettRNASummary}{GRangesList}(x)

\S4method{gettRNASummary}{GRanges}(x)
}
\arguments{
\item{x}{a GRanges or a GRangesList object. All elements have to pass the
\code{istRNAGRanges} test.}
}
\value{
a DataFrame object
}
\description{
\code{gettRNASummary} prepares a DataFrame with the aggregated features
of tRNAs from a GRanges object. Logical values are converted to numeric
values.
}
\examples{
data("gr", package = "tRNA")
gettRNASummary(gr)
}
