% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNA-checks.R
\name{istRNAGRanges}
\alias{istRNAGRanges}
\alias{istRNAGRanges,GRanges-method}
\title{tRNA compatibility check}
\usage{
istRNAGRanges(x)

\S4method{istRNAGRanges}{GRanges}(x)
}
\arguments{
\item{x}{the \code{GRanges} object to test for compatibility.}
}
\value{
a logical value
}
\description{
\code{istRNAGRanges} checks whether a GRanges object contains the
information expected for a tRNA result. This is used internally to ensure the
the required data is present in the input.
}
\examples{
data("gr", package = "tRNA")
istRNAGRanges(gr)
}
