% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNA.R
\docType{data}
\name{tRNA-data}
\alias{tRNA-data}
\alias{gr}
\alias{tRNA-data}
\alias{gr_human}
\alias{tRNA-data}
\alias{gr_human2}
\alias{tRNA-data}
\alias{gr_eco}
\title{tRNA example data}
\format{object of class \code{GRanges}}
\usage{
tRNA

gr_human

gr_human2

gr_eco
}
\description{
Example data for using the tRNA package
}
\keyword{datasets}
\keyword{tRNA}
