% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNA.R
\docType{data}
\name{tRNA-data}
\alias{tRNA-data}
\alias{gr}
\alias{gr_human}
\alias{gr_human2}
\alias{gr_eco}
\title{tRNA example data}
\format{
object of class \code{GRanges}

An object of class \code{GRanges} of length 596.

An object of class \code{GRanges} of length 631.

An object of class \code{GRanges} of length 89.
}
\usage{
data(gr)

data(gr_human)

data(gr_human2)

data(gr_eco)
}
\description{
Example data for using the tRNA package
}
\keyword{datasets}
\keyword{tRNA}
